+++ logic/src/commonTest/kotlin/com/wire/kalium/logic/data/user/UserRepositoryTest.kt
@@ -522,7 +522,7 @@ class UserRepositoryTest {
             given(selfApi)
                 .suspendFunction(selfApi::getSelfInfo)
                 .whenInvoked()
-                .thenReturn(NetworkResponse.Success(TestUser.USER_DTO.copy(deleted = true), mapOf(), 200))
+                .thenReturn(NetworkResponse.Success(TestUser.SELF_USER_DTO.copy(deleted = true), mapOf(), 200))
         }
 
         fun withSuccessfulGetMultipleUsersApiRequest(result: ListUsersDTO) = apply {
@@ -550,6 +550,6 @@ class UserRepositoryTest {
     }
 
     private companion object {
-        val SELF_USER = TestUser.USER_DTO
+        val SELF_USER = TestUser.SELF_USER_DTO
     }
 }
